Commit PC 31/3

+++ src/main/java/com/MarkusE/Wigell_Travel_API/service/CustomerService.java
@@ -69,6 +69,10 @@ public class CustomerService {
         Address address = addressRepo.findById(addressId)
                 .orElseThrow();
 
+        if (!address.getCustomer().getId().equals(customerId)) {
+            throw new RuntimeException("Address does not belong to this customer");
+        }
+
         addressRepo.delete(address);
     }
 }
